refactor(probe): migrate tcp to per-protocol plugin
