+++ tools/probe/tests/test_probe_protocol_coverage.py
@@ -38,7 +38,6 @@ from tools.probe.engine.protocols import PROTOCOL_REGISTRY, registered_plugins
 # seed and asserts full coverage directly.
 KNOWN_UNMIGRATED: frozenset[str] = frozenset(
     {
-        "tcp",
         "bgp",
         "rip",
         "ospf",
